// apps/git-client/src/generated/RebasePlanEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RebasePlanAction } from "./RebasePlanAction";

export type RebasePlanEntry = { oid: string, subject: string, parents: Array<string>, action: RebasePlanAction, message: string | null, published: boolean, mergeCommit: boolean, };